✅ 4Sum | Medium | C++ | Runtime: 521 ms (5.00%) | Memory: 83.7 MB | 2026-09-01

+++ leetcode/medium/0018-4sum/solution.cpp
@@ -3,13 +3,13 @@ public:
     vector<vector<int>> fourSum(vector<int>& nums, int target) {
         if(nums.size() < 4) return {};
         set<vector<int>> ans;
-        unordered_map<int,int> mp;
+        unordered_map<long long,long long> mp;
         for(int i = 0; i < nums.size(); i++) mp[nums[i]] =i;
         for(int i = 0; i < nums.size()-3; i++){
             for(int j = i+1; j < nums.size()-2; j++){
-                int newtarget = target - (nums[i]+nums[j]);
+                long long newtarget = (long long)target - (nums[i]+nums[j]);
                 for(int k = j+1; k < nums.size() ; k++){
-                    int com = newtarget - nums[k];
+                    long long com = newtarget - nums[k];
                     if(mp.count(com) && mp[com] != k && mp[com] > k){
                         vector<int> temp = {nums[i],nums[j],nums[k],nums[mp[com]]};
                         sort(temp.begin(),temp.end());
